Hopefully the branch has proved ok, this commit represents the progress made with this wrapper in tandem with the java-based archetype editor, with more methods available that the editor required.
Most of these methods have been collated into a class C_OBJECT_ACCESS with a few minor changes in other classes. The updated .def .c .h and .lib files for these changes are included also.

Some expansion on populating the hash tables is seen within the C_ARCHETYPE_NODE_POPULATOR class but whether this is the ideal solution is not known as yet, but as it is, works fine.