ADL Workbench can now be built on Mac OS X PowerPC with EiffelStudio 5.7.64380, installed according to the instructions at http://eiffelsoftware.origo.ethz.ch/index.php/MacPorts_Gtk_Installation.
It would have already worked on an Intel Mac, according to Manu, but the PowerPC version of gcc's cannot link with a precompiled vision2 library. We therefore now have two targets: adl_workbench for no precompiled library, and adl_workbench_precompile. Windows uses the latter, so the .msi installer project has to pick adl_workbench.exe from a different directory.

Because the SConstruct now specifies a target within the .ecf, it has exposed a bug in Eiffel.py, which was wrongly appending .exe to the target name (i.e., it was wrongly generating "adl_workbenchprecompile.exe"). This is now fixed.