Fix a problem with Unicode support. STRING_UTILITIES.utf8 was assuming that the given string is either UTF-8 or ASCII. This assumption fails if the string is actually Latin-1 (commonly used by many European languages).
For example, given "için", UC_UTF8_ROUTINES.valid_utf8 interprets the first byte as a whole character, but 'ç' (ordinal value 231) is interpreted as a three-byte character. Obviously this is wrong, and not surprisingly it decides it has an invalid encoding when it examines the next two bytes. This caused a precondition violation or, if assertions were disabled, the four-character string was wrongly displayed as an 'i' followed by a garbage character.

The fix is to check whether the given string is valid UTF-8. If it is not, then it is not converted. This seems to be working fine with some sample English, Turkish and Farsi text. It sounds like an unreliable hack, but according to http://czyborra.com/utf/ it may be the right thing to do: "You can detect that you are dealing with UTF-8 input with high probability if you see ... valid UTF-8 multibyte characters since it is very unlikely that they accidentally appear in Latin1 text. You usually don't place a Latin1 symbol after an accented capital letter or a whole row of them after an accented small letter."