Revert the last change to STRING_UTILITIES.utf8, to avoid the performance hit. It now explicitly requires that the given string be UTF-8 (and therefore, by implication, ASCII). Latin-1 is rejected.
Instead, on ADL Workbench's Test page, add a "UTF-8" test that verifies that each .adl file is UTF-8. I inserted the "UTF-8" test as the first to be executed, since it seems pointless parsing, etc. if the file isn't even in the correct format. If a file fails this test, its contents, assumed to be Latin-1, are converted to UTF-8 and written back out.

Note that conversion to UTF-8 uses UC_UTF8_ROUTINES.to_utf8, which calls some expensive preconditions in a tight loop, so it is extremely slow when run with preconditions enabled. To avoid this, I've disabled preconditions on UC_UTF8_ROUTINES.