Fix SHARED_RESOURCES.application_startup_directory and SHARED_RESOURCES.application_full_path, which were returning the application's current working directory rather than the start-up directory. This worked only if the user happened to start the application from its own directory (e.g., by double-clicking it in Windows Explorer); otherwise, it returned the wrong path.
These functions now always return the true path to the application, with one exception. If the application is in a directory of the form "p/EIFGENs/t/[WF]_code", then it is assumed to be the workbench or finalised directory built by an EiffelStudio project. So that the application can run, without needing copies of icons and various other files in the F_code or W_code directory, the functions go up three directory levels to return "p" (the project directory).