Extensive refactoring, renaming etc. - classes previously called CONSTRAINT_MODEL_XXX are now called C_XXX, which is in line with all the existing C_XXX classes   (i.e. this is the 'constraint model' - e C_OBJECT, ARCHETYPE_CONSTRAINT classes and so one on). This is partially as a   convenience, but also is probably clarifying... - inspired by recent Eiffel user group postings, visitor classes have had some renaming to make it more obvious to the outside   observer that that is what they are - e.g. SERIALISABLE has been renamed to VISITABLE; a C_VISITOR class has been created to   properly model the visitor concept for C_XXX object hierarchies; this will simplify some code in ARCHETYPE, in readiness for   the validation needed for specialisation. - the class C_SERIALISER_MGR is now redundant and no longer in use. - GUI_TEST_ARCHETYPE test routines have been simplified based on a more disciplined use of archetype_compiler. - no file renaming has been done to match class renames; probably this is not worthwhile at this stage.
All archetypes appear to run normally through the tests in the test view, but some proper validation needs to 
be done to be sure of this. The renames and other changes done here should probably be exported to the TRUNK. 
Note that no semantic changes have yet occurred, however, some functions used by the Archetype Editor, notably
ARCHETYPE_COMPILER.create_archetype and specialise_archetype don't yet work now - so a DLL built from this 
code won't do everything it needs to yet. Nevertheless, it should be tested with the Archetype Editor, to 
get the latter up to date.