AWB-13: Fix precondition violation "has_language" in ARCHETYPE.logical_paths. This can be replicated by following the steps given by the scenario test case: https://projects.oceanehr.com/confluence/display/TTL/View+in+Other+Languages
The violation was not occurring in the TRUNK. It was caused by ARCHETYPE_COMPILER.parse_archetype which, now that it is caching its parse results, was failing to ensure that the current language was correctly set. The error therefore occurred only after selecting an archetype a second time, following a change of language.