These additions now compute the 'inherited subtree list' of an archetype (see routine of that name in ARCHETYPE), which is the list of C_OBJECTs / paths to be removed from a current flat-style specialised archetype, along with the inherited ontology stuff, to produce a differential style archetype source.
The call is currently made as a new test case, via a new routine in ARCH_CONTEXT. This is by no means definitive, and may
only last until we have converted all the repositories (although it will need to remain a feature somewhere in the tool
for local users). For now, there is an extra test column, and this subtree list is computed on specialised archetypes.
Further work is required to perform removal of the found subtrees, which should not be hard.

To do this, a new generic iterator for C_XXX structures was written. This is perhaps slight overkill, but is more elegant code
and more comprehensible (IMO) than doing it all inline in the ARCHETYPE class. This generic iterator is used only for C_XXX
structures when the visitor pattern is not being used; note that it will need its do_all routine implemented if it becomes
necessary.