Various changes to make ADL workbench treat .adl files as 'flat' files and create .adls files to contain differential form source. - parsing looks for .adls file and tries to do something with it if found - the save operations on ARCHETYPE_COMPILER cause saving of both .adl and .adls, but only .adls for specialised files currently
Saving currently fails since serialising the differential form archetypes created by the ARCHETYPE.convert_to_differential operation fails due to codes mentioned in constraints that come from parents, i.e. are not in the ontology of the current archetype, but in some parent. Currently, the serialiser and a fair bit of the validation code assume that all codes found in an archetype are in the local ontology. We have to decide whether to make it search a flattened ontology, or all parent ontologies, or not to bother when serialising differential archetypes (i.e. just put a comment like "-- from parent") for each such code.

There will most likely be some moving around of the functionality I have currently created between the classes ARCHETYPE_COMPILER, ARCH_CONTEXT, and ARCH_REP_ARCHETYPE.