EDT-30: In Archetype Editor, deleting the references from the Description tab did not work because RESOURCE_DESCRIPTION.add_other_detail requires that the value be non-empty, but RESOURCE_DESCRIPTION provides no procedure for removing an item from the other_details list. To fix this, Archetype Editor now simply ignores the precondition, removing its check for an empty string before calling add_other_detail; this means that the reference is replaced with an empty string, as in the following example:
	other_details = <
		["references"] = <"">
	>

This works because Archetype Editor is deployed with a build of adl_dotnet_lib.dll that has discarded assertions. This is ok as a short-term fix for the bug. In the longer term, when we are able to get a working build of adl_dotnet_lib.dll, RESOURCE_DESCRIPTION must provide a way of removing an item from the other_details list. There are two obvious solutions: either add a remove_other_detail procedure, else change add_other_detail to accept an empty string and remove the item. I've opted for the latter fix, because the current Archetype Editor call will work in the desired manner as soon as it gets a new build of adl_dotnet_lib.dll.