EDT-145: Update a link on the project page, broken by the previous revision.
The previous revision also broke the SConstruct's build of the adl_parser, because a non-existent ECF target was specified now that the build target and the ECF target have different names. (Note that the Aug2007 tag's SConstruct remains broken, because the Eiffel builder was incapable of handling this difference until later in August, after revision 310 on which this tag is based.)

Also, this showed up a pre-existing error in the SConstruct: the unmanaged DLL was not specified as a side-effect of building the .NET DLL, so if it did not already happen to exist then SCons did not know how to create it for installation in the oe_distrib directory.