AWB-65: In both the TRUNK and BRANCHES/specialisation, set version number 1.4.1.0. Previously, ADL Workbench did not have a version number, although informally the ADL version 1.4 seemed to be used informally for this purpose; and the ADL Parser .NET DLL was at version 0.0.0.2; so progressing to 1.4.1 seems consistent with what has gone before. The version number now appears in ADL Workbench's splash screen and About box.
Make the version number, and other version information, appear in the Windows Explorer Properties by setting them in ADL Workbench's resource file. We should do the same for the ADL Parser DLL too, but the resource file technique does not work. Eiffel Software currently provides no way to do this for .NET assemblies; even setting an assembly-level ASSEMBLY_FILE_VERSION_ATTRIBUTE does not work, despite the fact that this is apparently how it is achieved in C# and Visual Basic. So for now, the version information will not be available in Windows Explorer for the .NET DLL.

The final part of the version number (i.e., the zero in 1.4.1.0) is replaced for a release build by the current Subversion revision number. This is done by the SConstruct, using similar techniques to those implemented for Archetype Editor. Unlike Archetype Editor, which has a pre-build event that sets the version information in a non-version-controlled source file that it creates specifically for this purpose, the relevant ADL Workbench and Parser files are updated by the SConstruct. I chose not to use an EiffelStudio pre-build task for this purpose for two reasons: (1) pre-build tasks do not work at all in EiffelStudio at the moment; and (2) ADL Workbench is cross-platform, so a Windows batch file is not suitable. The SConstruct avoids this problem because (1) it works, and (2) SCons is cross-platform. The SConstruct is a good place to do it also because developers rarely use it during day-to-day work, so anything built within EiffelStudio will be identifiable by the zero revision number; but proper release builds will be identifiable by the non-zero revision number auto-generated by the SConstruct. The disadvantage is that the code to do this is more complex: not only does it have to search and replace the revision number within the relevant files; it also has to take care to restore each file to its prior state after the ADL Workbench and Parser have been built, recovering gracefully if the build process aborted abnormally.

We could improve this by completely generating the files within the SConstruct: the ADL Workbench resource file would be easy; and OPENEHR_VERSION might have an auto-generated descendant, non-version-controlled and containing only the revision number. This would also reduce the maintenance burden of updating the manually-updated parts of the version number in multiple files. The SConstruct would still have to auto-edit adl_parser.ecf, however; perhaps we could auto-generate a class containing the necessary assembly-level attributes (ASSEMBLY_VERSION_ATTRIBUTE, etc.), in a manner very similar to what Archetype Editor is doing: then we would omit the version info completely from adl_parser.ecf!

The revision number is not set in the NSIS script's version number. It therefore reports 1.4.1.0. Maybe we should do this too, one day, but it seems low-priority to me. Also, the Mac Info.plist file contains <plist version="1.0">, which I suspect should be changed to 1.4.

Update the copyright date to 2008, because we have almost reached the new year!