AWB-72: Enable building of installers for Tiger and Leopard, PowerPC and Intel.
There are two possible overall approaches to the problem of incompatible library dependencies (GTK, etc.): either bundle all of the necessary dependencies up within the application package so as to isolate it from the rest of the machine; else rely on the user installing all of the necessary dependencies. The former was the approach taken by our first attempts in 2006, imitating EiffelStudio 5.6 and earlier. The latter is the approach taken by the current EiffelStudio, and this is the solution now adopted for ADL Workbench.

The last release in April was actually a hybrid of the two approaches, still using some 3rd-party GTK libraries (the same as EiffelStudio 5.7). We got these libraries as Subversion externals. This was a nuisance, because it slowed down Subversion updates. One benefit of adopting the pure second approach is that we no longer get these externals. (For convenience, I'm removing these external references from all branches.) Another benefit is that the launcher.sh script no longer has to initialise variables and settings files for these libraries.

The launcher.sh script is further simplified because of a big improvement in Leopard, which now launches the X server automatically when adl_workbench is run. Unfortunately, for backwards compatibility, launcher.sh still has to detect if it is running on Tiger and, if so, it makes sure that $DISPLAY is initialised and launches the X server. (Note that it dispenses with the old complicated logic for calculating $DISPLAY, simply setting it to :0.0 if it is not yet set; this seems to be all that is needed for running ADL Workbench as a simple desktop application.) See http://eiffelstudio.origo.ethz.ch/node/33 for a discussion of issues in getting EiffelStudio running in Leopard; the same issues arose for ADL Workbench.

There are a couple of other improvements in launcher.sh:
1. It always opens the Console application. It used to do so only if there was an error starting adl_workbench. Hopefully users won't find this annoying, but it should help us enormously when troubleshooting installation difficulties.
2. Before creating symbolic links to ADL Workbench in the user's Application Support folder (which works around ADL Workbench writing to its own application directory), launcher.sh unlinks any symbolic links that it already finds there. This will ensure removal of, for example, the old GTK links left lying around from the April release.

An enhancement to the installer is that it now automatically runs ADL Workbench when installation has completed. This is consistent with the Windows installer.

Both application packages (the application and the installer) now have version and copyright information, visible when you Get Info in Finder. Consistent with adl_workbench itself, this is currently 1.4.1.x, where x is automatically set by the SConstruct to the Subversion revision number. (It now sets the revision number in the Windows installer too.)

The SConstruct's logic has changed for backing up each file into which it injects the revision number. It used to remove the backup if it already existed before renaming the original to the backup; but if the build was aborted (which is quite a common occurrence), then this meant that the original file will be lost during the next build, and unless the programmer notices what has happened then a non-zero revision number is likely to be committed to Subversion shortly thereafter. To minimise this likelihood, it no longer renames the original to the backup if the backup already exists. This tends to preserve the original file, even after an aborted build.

The SConstruct now wraps the installer (.pkg) directory up into a standard Mac disk image (.dmg) file. This .dmg file is what we will commit to the oe_distrib Subversion repository: we no longer have to create a .zip or .dmg manually. The names of the .pkg and the .dmg are the same, apart from their extensions. Because the build is platform-specific, the name indicates the platform, e.g., "ADL Workbench for Tiger powerpc.dmg" or "ADL Workbench for Leopard i386.dmg". The oe_distrib repository will therefore contain a .dmg for each platform, from which the user will have to choose.

Leopard's packagemaker utility has a new command line interface, but the SConstruct is still using the old "-build" options (which are still supported) for backward compatibility with Tiger. Unfortunately, Leopard's packagemaker has an incompatibility: it copies all of the ".svn" directories into the installation package. To avoid this, rather than the old approach of copying adl_workbench, news.txt, the icons, etc. into "ADL Workbench.app" within the Subversion source tree, the SConstruct now copies everything into a temporary directory within the local oe_distrib. (It doesn't copy anything that starts with ".", of course.) As well as working around the packagemaker bug, this avoids cluttering the local source directories with the copied files; it allows the app's "bin" directory to be removed, placing its files directly under "Resources"; and it makes it easy to delete everything to ensure that nothing left over from an old build gets into the package that doesn't belong there.

This new approach caused a new problem, however. When testing the installer, I found that it refused to install the newly-built "ADL Workbench.app" into the Applications folder. Inspection of the install logs revealed that it was "diverting" to the temporary "ADL Workbench.app" within oe_distrib. I haven't discovered the rationale for this crazy "diverting" behaviour, but as a workaround the SConstruct removes the temporary "ADL Workbench.app" within oe_distrib, as soon as the installer package has been built. This should avoid the risk of another developer wasting hours, as I did, trying to figure out why the installer wasn't working!

This commit has been tested on Leopard i386 and on Tiger powerpc. Both machines have installed Xcode, X11, MacPorts and EiffelStudio 6.1.