AWB-71: Remove the HTML column from the Test page. Instead, generate HTML for the currently selected archetype when the user selects the Tools | Export HTML menu option.
The HTML file is generated under the configured HTML Export directory, concatenated with the archetype's ontological path. All necessary intermediate directories are created automatically.

The HTML Export directory is now initialised, during program start-up, to be an "html" directory immediately below the repository path. This is not done the first time the program is run, however, in order to give the user an opportunity to set the repository path correctly; otherwise it would be wrongly initialised to the application start-up directory.