AWB-71: Under the Tools menu, the Export HTML and Clean Generated Files options now ignore the node selected in the repository explorer. They now always operate on the whole repository, which is almost certainly what the user would always want. (Note that, if the user wants HTML for a single archetype, this functionality is still available via File | Save As.)
On choosing to export HTML, the user must answer a dialog that gives the option to build each archetype before exporting it. This is the default behaviour, and it ensures that all valid archetypes are exported; but if the user declines this option, then only the archetypes that have previously been built can be exported.

The logic for exporting to HTML is now in ARCHETYPE_COMPILER. This allows it to be integrated into the build process, rather than having to build all before exporting them all. The user can interrupt the export using the same mechanism that is available when building.

Also, fix the restoring of the main window's splitter positions. They were being restored from the configuration settings before the window was shown, but this does not seem to work so now we make sure to restore them after showing the window.