AWB-13: Fix grafting an ad hoc archetype into the repository tree, where (a) either the archetype's file name indicated that it was not specialised, or else the specialisation parent was not present in the repository; and (b) the last archetype in the repository for the ad hoc archetype's RM entity happened to be a specialisation. The ad hoc archetype was grafted under the parent archetype of the specialisation. It is now grafted correctly under the RM entity folder.
Also, update the main window's status panel after attempting to open an ad hoc archetype. If the file name is invalid, display an error, similar to repository archetypes. This is likely to happen if, say, the file name is "xx.adl".

To help draw the user's attention, change the background colour of the status panel depending on whether there is an error.