AWB-46: After displaying two archetypes, successively, in a non-English language, the language combo box showed "en" despite the fact that the archetype was being shown the other language.
Also, fix a precondition violation when the current language was non-English for an English-only archetype. This occurred because the current language was not being reset on changing archetype. The bug was discovered at step 4 of https://projects.oceanehr.com/confluence/display/TTL/View+in+Other+Languages; i.e., view a multi-language archetype in the non-English language; then view an English-only archetype; then repeat both steps again, and the crash happened on viewing the English-only archetype the second time.