AWB-15: At Thomas's request, all .adls files in the repository are now deleted automatically on start-up for the first time, to ensure the effects of previous bugs are erased. For simplicity (because this is a temporary change), this is done whenever news.txt has changed; so it will only work if we make sure that we change news.txt between releases. Another caveat is that it will only work on the configured repository paths, so if the user switches repository paths then they could pick up some old .adls files: they should run the clean menu option manually.
Also, change how the toolbar and menu icons are loaded. Previously they were loaded via EiffelBuild from the "./icons" subdirectory, but this can cause ADL Workbench to raise an exception during start-up if the current working directory is not the application directory. EiffelBuild's own "display" window fails to show them anyway, because the EiffelBuild project is not in the directory above the icons directory, so we are missing out on that nice design-time WYSIWIG effect in EiffelBuild. Instead, the toolbar and menu icons are loaded using the same mechanism that we are using for all of the other pixmaps, which is immune from variations in the current working directory.