AWB-15: Fix ARCHETYPE_VALIDATOR.validate_found_codes, which was not checking that at-codes were at the correct specialisation level. This caused a Void target call in DIFFERENTIAL_ARCHETYPE_ONTOLOGY.has_term_code, which is now protected against; likewise in has_constraint_code.
Also, add invariants to AUTHORED_RESOURCE_VALIDATOR that target and target_descriptor are attached. There was already an implicit assumption in ARCHETYPE_VALIDATOR.ontology that this would be so.

Also, because ARCHETYPE_VALIDATOR.ontology could never be Void, there was no point in ARCHETYPE_VALIDATOR.validate_basics checking whether it is Void. (This was explicitly guaranteed by a run-time invariant in ARCHETYPE, and has now been proven by instead marking it attached at compile time. This use of a compile-time attachment marker had a knock-on effect into ADL_ENGINE.parse, which creates ontology objects; resolving this has also simplified the convoluted logic for initialising the parsed archetype's original language and translations.)