AWB-83: Fix a void target call in MAIN_WINDOW.populate_archetype_id. It occurred if a folder node became selected in the repository tree while a previously-selected archetype was still in the process of being displayed. The flow of logic was wrongly assuming that the archetype would still be selected.
This tended to show up only when the selected archetype was changed programmatically (e.g., after clicking the Back or Forward buttons or double-clicking in the Errors list), as opposed to being manually selected by the user in the repository tree. This was because, immediately after programmatic selection in the directory and in the tree, it displayed it; and then, after a brief delay, the selection event triggered it to be displayed again. As well as increasing the likelihood of trying to display the wrong thing, double-displaying the archetype caused some flickering. To fix this, the archetype is no longer displayed immediately after selection; instead, it relies on the selection event. This means that there is now a slight delay after clicking the Back or Forward buttons, etc., which is more consistent with what happens when the user selects an archetype directly in the tree. It gives the additional benefit of being more responsive if the user goes back and forth by holding down the Alt+Left and Alt+Right keys, because it is no longer unnecessarily spending time displaying archetypes that the user is simply flitting past.