AWB-14: Add a new tab page for viewing the archetype's flat text: so the "Source" tab now has a "Flat Text" tab to its right. The Source tab now never shows the flat text; instead, if there is no source file, it just shows a message saying that there is no source file. (This makes it easy to know whether we are looking at the source or the flat text: with the old arrangement, I was constantly getting confused about which I was looking at. Also, the line numbers now match up properly with the parser error messages; before, they were always off by one in the flat text.)
Doing this has made it clear that there was a minor bug: on first selecting an archetype that has no source file, the archetype would be compiled, generating the source file; but it still said that there was no source file. This is now fixed by redisplaying the source after compiling.