AWB-84: On the Slots tab, indicate the number of slots in the current archetype and the number of other archetypes that use the current archetype, e.g. "Slots (0/12)".
In order to do this within the controller classes rather than within MAIN_WINDOW, GUI_USED_BY_MAP_CONTROL has been merged into GUI_SLOT_MAP_CONTROL. This refactoring is a considerable simplification.

Also, to enhance keyboard access and for consistency with the Errors grid, the Slots tree widgets now respond to the Enter key by navigating to the relevant archetype, in the same way as a double-click.