Corrected the logic in ARCHETYPE_FLATTENER.node_graft, and also added attributes paths back into OG_OBJECT_NODE.generate_all_paths, because otherwise ARCHETYPE.has_path was incorrectly responding negative to attribute paths, and objects with no predicate, when they are in fact present.
Further testing required, but the output for the lab archetypes now looks correct, although there are some anomalies which are probably due to the archetypes - e.g. repeated occurrence of use_node ELEMENT under at0011.1 node in the lipids archetype. This version is nevertheless far closer to correct than the previous one.