EDT-192: In DATE_TIME_ROUTINES, valid_time_constraint_patterns and valid_date_time_constraint_patterns were inconsistent. One extra pattern has been added to each.
In particular, the partial date-time pattern "yyyy-mm-ddTHH:??:??" was missing from valid_date_time_constraint_patterns and is needed because Archetype Editor generates it. On reopening an archetype that Archetype Editor had just saved with this pattern, the parser would reject it.

Because this fix is needed immediately by Archetype Editor, it is being applied to both the TRUNK and to TAGS/Aug2007, from which the live live version of the ADL Parser DLL is built. To assist building, the current TRUNK SConstruct has been copied to TAGS/Aug2007; this is ok because we only need to build the DLL, and there are no inconsistencies that affect it. Note, however, that this has resulted in the regeneration of the gelex and geyacc classes within TAGS/Aug2007, although as expected the changes do not appear to be significant.