AWB-117: Because of revision 619, which made sure that an invalid .adl file was re-parsed after it had changed, archetypes that had already been successfully compiled were recompiled whenever the user selected them in the repository tree. This is fixed by a major reworking of the compilation process: * The decision whether to recompile an archetype selected in the repository tree was previously split between MAIN_WINDOW.parse_archetype and ARCHETYPE_COMPILER.build_archetype. The decision now rests entirely with the latter. * The decision is now based on whether it is valid, not on whether it parsed successfully. This is a more correct way of fixing the specific problem mentioned in AWB-117, i.e. after the user edits an archetype with a validity error, it gets re-parsed correctly. * As a consequence of this, the archetype's lineage is considered for recompilation. Previously, if the selected archetype's parent was changed on disk, the selected archetype would not be recompiled. This should now be fixed, although this scenario needs testing. * Side-effects on timestamp attributes in ARCHETYPE_FILE_REPOSITORY_IMP and ARCH_REP_ARCHETYPE have been eliminated in the reading of the .adl and .adls files. These were a major cause of confusion in knowing whether the file had changed on disk. Reading the files is now done by explicit commands, not as a side-effect of asking for the text. * As a consequence of this, the files are now read less frequently. Previously they were read each time the user switched between the Differential and Flat views, due to the side-effect. They are now only read during parsing. * ARCH_REP_ARCHETYPE used to have separate queries to check whether the .adl or .adls files had changed with respect to the timestamp attributes, of which there ere also two. This is reduced to one query and one timestamp attribute. The latter represents either the .adl or .adls file, whichever was read or saved most recently. It is now also set on saving; previously it was only set on reading. * In order to know which file the timestamp refers to, ARCH_REP_ARCHETYPE.full_path is set to either the .adl or the .adls file whenever the timestamp is set on reading or saving. * Having only one timestamp attribute per ARCH_REP_ARCHETYPE makes the test for whether the file is out of date very simple, whereas the previous arrangement would have required careful logic to compare against one or other of the old timestamp attributes. * As a consequence of this, the files are now accessed less frequently to decide whether they are out of date. ARCHETYPE_COMPILER.build_archetype is now the only place where this is called.
AWB-137: Previously, ARCH_REP_ARCHETYPE read the flat text from the full_path rather than the .adl file. Usually this happened to be correct, because the repository is populated by searching for .adl files which were then set as the full_path. But in future we will be populating by searching for .adls files. Also, there was already a bug when opening an ad-hoc .adls file: it always displayed the .adls source in the Flat view, because in this scenario full_path was the .adls file. This is fixed by reading the flat text correctly from the .adl file. Because we have eliminated the side-effect of reading the file on querying the text (as mentioned above), the job of ensuring that the .adl file actually does get read now occurs in ARCH_REP_ARCHETYPE.set_archetype_differential, after the flattened archetype is generated. It is done by reading the .adl file, much as before, but without the side-effect. If there is no .adl file, then the flat text is empty: an appropriate message now appears in the Flat view's Source box, similar to the one that was already being shown if there is no .adls file. This resolves the bug with opening an ad-hoc .adls file. It might be better, however, rather than reading the .adl file, to set the flat text by serialising the flattened archetype.