The target passed to the SCons Eiffel builder used to be the base name of the main executable or .dll that was to be built, but this "target" was confusingly different from the ECF project "target". The builder now takes the ECF target instead. As well as avoiding confusion, this simplifies the SConstruct's use of the builder, because it now has only one "target" to worry about and longer has to append '-target' to ECFLAGS. The Eiffel builder works out the path to the files it is building by reading the .ecf file. This also allows it to emit automatically the additional build targets of several kinds of project: .NET assemblies, shared libraries, and precompiled libraries, which further simplifies the SConstruct.
The SConstruct also fixes unnecessary rebuilds that could be done when alternating between simply building adl_workbench and adl_parser or building installers for them. This was because the SConstruct was setting the Subversion revision only when building installers; SCons was smart enough to recognise that the build conditions had changed, so it rebuilt them, and of course they were different! It now sets the revision either way, which is more consistent and avoids the unnecessary rebuilds.