AWB-146: Compiling in EiffelStudio 6.3 release candidate found 3 cases where an assignment attempt was being made to attached local variables, all in ARCHETYPE_VALIDATOR.specialised_node_validate.
I've fixed one of these by using an object test instead of the assignment attempt. (I also added a "check" assertion in case the object test ever returns false; the assignment attempt probably should have been doing this anyway. Inspection of how the local variable was subsequently used showed a couple of missing preconditions in C_ATTRIBUTE, so I've added these too.)

I haven't fixed the other two of these errors, because the logic is quite complicated, with the variable being assigned twice. Object test is inconvenient in such situations. I think the best solution would be to extract some of the code from specialised_node_validate out to another routine. I'm leaving these two assignment attempts to be fixed later, but they will have to be fixed soon because they don't compile in 6.3.

I also removed all uses of the "!" attachment marker from object tests in these two classes. The marker was needed in object tests in EiffelStudio 6.1, but the Eiffel ECMA committee reversed its decision on this, so beginning with EiffelStudio 6.2 the markers are not required.