EDT-458: Fix a bug on saving an ADL file in .NET: wherever a carriage return was already present in the in-memory serialisation of the archetype, the ADL file was saved with another carriage return inserted. The extra carriage return is inserted by the .NET implementation of PLAIN_TEXT_FILE.put_string, which inserts a carriage return before each linefeed that it finds in the given string; it seems to be assuming that carriage returns will not be present yet. This bug was apparent when saving an ADL file in Archetype Editor if the Purpose, Use, Misuse or References fields contained newlines: for each newline that the user entered, the ADL file contained two carriage returns followed by a linefeed.
This bug has been benign in the past, because on reading the ADL file the carriage returns are all removed. A user who opened the ADL file in ADL Workbench or Archetype Editor would see only the proper newlines. But another bug introduced recently in Archetype Editor exacerbated the problem, causing an extra line to be inserted every time the archetype was saved. Although the latter bug has been fixed too, in the course of tracking it down I fixed the insertion of the extra carriage return. Therefore, I'm committing the fix to BRANCHES/specialisation, as well as to TAGS/Aug2007 which contains the source code for the current public build of the ADL Parser DLL used by Archetype Editor.

The fix is to replace each carriage return-linefeed pair in the string with a simple linefeed prior to saving. Possibly a better fix would be to remove every carriage return; although I can't see any problem with that solution, it is conceivable that it might cause some other problem. As we are about to do a release of Archetype Editor, I chose the least risky fix.

While doing this fix in FILE_CONTEXT, I also removed some constant declarations that were duplicating definitions that already existed in the Gobo library. The Gobo constants are now being used. I also upgraded the source code for FILE_CONTEXT in TAGS/Aug2007 to be as close as possible to the same class in BRANCHES/specialisation.