Fix the way quoting is done on C_STRINGs and also correct the implementation (no change to the outcome) for DT_PRIMITIVE_OBJECTs (the objects of dadl strings). The design idea is: quoting is a concept that only applies to serialisation, and generally is specific to the medium. So... we treat ADL serialisation as one kind of serialisation which means that it is the serialiser classes that have to know about quoting - it should not be done from inside the originating objects that hold the data (after all, they generally don't want any quoting when being written to the screen, or other media, or might want different quoting for other reasons).
The result is somewhat messy, because both C_STRING and DT_PRIMITIVE_OBJECT and DT_PRIMITIVE_OBJECT_LIST do messy things in order to deal with the various types of their ultimate primitive payload objects (INTEGERs, STRINGs, SEQUENCE[STRING], etc). But for now, this is probably about the most correct approach until we rewrite some of this code.