AWB-158: Revision 801 was causing a void target call on the Tests page if the "Save ADLS" check box is off. The exception was happening during the running of the "Diff" test, which compares the source generated by "Save ADLS" and "Reparse". The exception is fixed by not running "Diff" if the source generated by "Save ADLS" is empty.
Note that the exception was caused by the fact that "Reparse" does not seem to regenerate the source: the new source string was void, causing the void target call. This drew my attention to a problem: "Reparse" doesn't seem to regenerate the source when even when "Save ADLS" has run successfully. The old and new source are the same string reference. It seems to me that the "Diff" test is completely meaningless, because it just seems to be testing that the original source is the same as itself, which is always true. To highlight this, I have added a postcondition to 'test_reparse_differential'. The postcondition always fails, so I have commented it out, with a FIXME comment.